# Fix ufunc performance degradation