@takagi 2018-12-22
@asi1024 Would you make the backport of this PR because I think of backporting #1905?