@beam2d 2019-01-09
It causes a performance degradation in #5958, so I created a backport.